Lend a credit for the HCI connection at core_init_rsp
Due to a race condition in AOSP, the core_conn_credits_ntf for the HCI
connection may be handled before the core_init_rsp (in which the
HCI conn id is set). Therefore the hci num_buff variable is set wrongfully
at 0.
To avoid it, a credit is lent at the core_init_rsp and taken back at the
reception of the next core_conn_credits_ntf.

Change-Id: I3adf5f552764958759ae74c63a155ac4e4d19d46