Avoid system lock when CLF has always data to send
In some specific cases the CLF may have a constant stream of
data or a the IRQ is stuck at high.

This patch changes the logic to give a slot to the write and exit
command every 3 reads at most.

Bug: 162814256
Bug: 162804529
Test: Nfc hal would not be locked when i2c bus is stuck
Merged-In: Id370dc6dfd68c8594442e3de5f6edae6450fa41a
Change-Id: Id370dc6dfd68c8594442e3de5f6edae6450fa41a