Allow timeouts when sending the MODE OFF command
When the HAL is closing, if the CLF is not responding at that time, we
can end up in a state where we don't receive a response to the
PROP_MODE_SET final command. In that case, we need the HAL to be closing
anyway.

Bug: 162814256
Bug: 162804529
Test: Nfc hal would not be locked when i2c bus is stuck
Merged-In: Ide7eebe31664d96d3e62054ac27cdbd7fb2f0b19
Change-Id: Ide7eebe31664d96d3e62054ac27cdbd7fb2f0b19